requirements.txt
blessed==1.14.2
PyYAML=3.12
